Updates as on 21/5/2018

+++ src/main/java/com/technocomp/emsapp/domain/Group.java
@@ -15,7 +15,6 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.Transient;
-import javax.validation.constraints.NotNull;
 import org.springframework.format.annotation.DateTimeFormat;
 
 /**
@@ -33,7 +32,7 @@ public class Group {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "groupName")
     private String groupName;
@@ -54,11 +53,11 @@ public class Group {
         this.dateAndTime = dateAndTime;
     }
     
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -77,7 +76,4 @@ public class Group {
     public void setGroupEmail(String groupEmail) {
         this.groupEmail = groupEmail;
     }
-    
-    
-    
 }
